move agent setting to common place
cc @roblourens
