+++ src/vs/workbench/contrib/chat/browser/chat.contribution.ts
@@ -81,7 +81,7 @@ import { QuickChatService } from './chatQuick.js';
 import { ChatQuotasService, IChatQuotasService } from '../common/chatQuotasService.js';
 import { ChatResponseAccessibleView } from './chatResponseAccessibleView.js';
 import { ChatEntitlementsService, ChatSetupContribution } from './chatSetup.js';
-import { IChatEntitlementsService } from '../common/chatEntitlementsService.js';
+import { ChatEntitlement, IChatEntitlementsService } from '../common/chatEntitlementsService.js';
 import { ChatVariablesService } from './chatVariables.js';
 import { ChatWidgetService } from './chatWidget.js';
 import { ChatCodeBlockContextProviderService } from './codeBlockContextProviderService.js';
@@ -100,6 +100,7 @@ import { registerChatToolActions } from './actions/chatToolActions.js';
 import { ChatStatusBarEntry } from './chatStatus.js';
 import product from '../../../../platform/product/common/product.js';
 import { ChatEditingNotebookFileSystemProviderContrib } from './chatEditing/chatEditingNotebookFileSystemProvider.js';
+import { Event } from '../../../../base/common/event.js';
 
 // Register configuration
 const configurationRegistry = Registry.as<IConfigurationRegistry>(ConfigurationExtensions.Configuration);
@@ -299,34 +300,39 @@ class ChatResolverContribution extends Disposable {
 	}
 }
 
-class ChatAgentSettingContribution implements IWorkbenchContribution {
+class ChatAgentSettingContribution extends Disposable implements IWorkbenchContribution {
 
 	static readonly ID = 'workbench.contrib.chatAgentSetting';
 
 	private registeredNode: IConfigurationNode | undefined;
 
 	constructor(
-		@IWorkbenchAssignmentService experimentService: IWorkbenchAssignmentService,
+		@IWorkbenchAssignmentService private readonly experimentService: IWorkbenchAssignmentService,
 		@IProductService private readonly productService: IProductService,
 		@IContextKeyService contextKeyService: IContextKeyService,
+		@IChatEntitlementsService private readonly entitlementService: IChatEntitlementsService,
 	) {
+		super();
+
 		if (this.productService.quality !== 'stable') {
-			this.registerSetting();
+			this.registerEnablementSetting();
 		}
 
 		const expDisabledKey = ChatContextKeys.Editing.agentModeDisallowed.bindTo(contextKeyService);
 		experimentService.getTreatment<boolean>('chatAgentEnabled').then(enabled => {
 			if (enabled) {
-				this.registerSetting();
+				this.registerEnablementSetting();
 				expDisabledKey.set(false);
 			} else if (enabled === false) {
 				this.deregisterSetting();
 				expDisabledKey.set(true);
 			}
 		});
+
+		this.registerMaxRequestsSetting();
 	}
 
-	private registerSetting() {
+	private registerEnablementSetting() {
 		if (this.registeredNode) {
 			return;
 		}
@@ -352,6 +358,34 @@ class ChatAgentSettingContribution implements IWorkbenchContribution {
 			configurationRegistry.deregisterConfigurations([this.registeredNode]);
 			this.registeredNode = undefined;
 		}
+	}
+
+	private registerMaxRequestsSetting(): void {
+		let lastNode: IConfigurationNode | undefined;
+		const registerMaxRequestsSetting = () => {
+			const treatmentId = this.entitlementService.entitlement === ChatEntitlement.Limited ?
+				'chatAgentMaxRequestsFree' :
+				'chatAgentMaxRequestsPro';
+			this.experimentService.getTreatment<number>(treatmentId).then(value => {
+				const defaultValue = value ?? (this.entitlementService.entitlement === ChatEntitlement.Limited ? 5 : 15);
+				const node: IConfigurationNode = {
+					id: 'chatSidebar',
+					title: nls.localize('interactiveSessionConfigurationTitle', "Chat"),
+					type: 'object',
+					properties: {
+						'chat.agent.maxRequests': {
+							type: 'number',
+							markdownDescription: nls.localize('chat.agent.maxRequests', "The maximum number of requests to allow Copilot Edits to use per-turn in agent mode. When the limit is reached, Copilot will ask the user to confirm that it should keep working. \n\n> **Note**: For users on the Copilot Free plan, note that each agent mode request currently uses one chat request."),
+							default: defaultValue,
+							tags: ['experimental']
+						},
+					}
+				};
+				configurationRegistry.updateConfigurations({ remove: lastNode ? [lastNode] : [], add: [node] });
+				lastNode = node;
+			});
+		};
+		this._register(Event.runAndSubscribe(Event.debounce(this.entitlementService.onDidChangeEntitlement, () => { }, 1000), () => registerMaxRequestsSetting()));
 	}
 }
