+++ src/vs/workbench/contrib/chat/browser/chatSetup.ts
@@ -64,11 +64,10 @@ import { ExtensionUrlHandlerOverrideRegistry } from '../../../services/extension
 import { IWorkspaceTrustRequestService } from '../../../../platform/workspace/common/workspaceTrust.js';
 import { toErrorMessage } from '../../../../base/common/errorMessage.js';
 import { StopWatch } from '../../../../base/common/stopwatch.js';
-import { IConfigurationRegistry, Extensions as ConfigurationExtensions, IConfigurationNode } from '../../../../platform/configuration/common/configurationRegistry.js';
+import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from '../../../../platform/configuration/common/configurationRegistry.js';
 import { IQuickInputService } from '../../../../platform/quickinput/common/quickInput.js';
 import { ILifecycleService } from '../../../services/lifecycle/common/lifecycle.js';
 import { equalsIgnoreCase } from '../../../../base/common/strings.js';
-import { IWorkbenchAssignmentService } from '../../../services/assignment/common/assignmentService.js';
 import { ChatEntitlement, IChatEntitlements, IChatEntitlementsService } from '../common/chatEntitlementsService.js';
 import { IStatusbarService } from '../../../services/statusbar/browser/statusbar.js';
 
@@ -164,8 +163,7 @@ export class ChatSetupContribution extends Disposable implements IWorkbenchContr
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@ICommandService private readonly commandService: ICommandService,
 		@ITelemetryService private readonly telemetryService: ITelemetryService,
-		@IWorkbenchAssignmentService private readonly experimentService: IWorkbenchAssignmentService,
-		@IChatEntitlementsService chatEntitlementsService: ChatEntitlementsService,
+		@IChatEntitlementsService chatEntitlementsService: ChatEntitlementsService
 	) {
 		super();
 
@@ -180,7 +178,6 @@ export class ChatSetupContribution extends Disposable implements IWorkbenchContr
 		this.registerChatWelcome(controller, context);
 		this.registerActions(context, requests);
 		this.registerUrlLinkHandler();
-		this.registerSetting(context);
 	}
 
 	private registerChatWelcome(controller: Lazy<ChatSetupController>, context: ChatSetupContext): void {
@@ -360,36 +357,6 @@ export class ChatSetupContribution extends Disposable implements IWorkbenchContr
 				return true;
 			}
 		}));
-	}
-
-	private registerSetting(context: ChatSetupContext): void {
-		const configurationRegistry = Registry.as<IConfigurationRegistry>(ConfigurationExtensions.Configuration);
-
-		let lastNode: IConfigurationNode | undefined;
-		const registerSetting = () => {
-			const treatmentId = context.state.entitlement === ChatEntitlement.Limited ?
-				'chatAgentMaxRequestsFree' :
-				'chatAgentMaxRequestsPro';
-			this.experimentService.getTreatment<number>(treatmentId).then(value => {
-				const defaultValue = value ?? (context.state.entitlement === ChatEntitlement.Limited ? 5 : 15);
-				const node: IConfigurationNode = {
-					id: 'chatSidebar',
-					title: localize('interactiveSessionConfigurationTitle', "Chat"),
-					type: 'object',
-					properties: {
-						'chat.agent.maxRequests': {
-							type: 'number',
-							markdownDescription: localize('chat.agent.maxRequests', "The maximum number of requests to allow Copilot Edits to use per-turn in agent mode. When the limit is reached, Copilot will ask the user to confirm that it should keep working. \n\n> **Note**: For users on the Copilot Free plan, note that each agent mode request currently uses one chat request."),
-							default: defaultValue,
-							tags: ['experimental']
-						},
-					}
-				};
-				configurationRegistry.updateConfigurations({ remove: lastNode ? [lastNode] : [], add: [node] });
-				lastNode = node;
-			});
-		};
-		this._register(Event.runAndSubscribe(Event.debounce(context.onDidChange, () => { }, 1000), () => registerSetting()));
 	}
 }
